+++ server/dfr3-service/src/main/java/edu/illinois/ncsa/incore/service/dfr3/daos/IMappingDAO.java
@@ -19,7 +19,7 @@ import java.util.Optional;
 public interface IMappingDAO {
     void initialize();
 
-    List<MappingSet> getMappingSets();
+    List<MappingSet> getMappingSets(String dataType);
 
     Optional<MappingSet> getMappingSetById(String id);